refactor: 태그 검색 쿼리 파라미터 타입 변경(dto -> String), 불필요한 코드 정리
