+++ src/main/java/com/windfall/global/exception/ErrorCode.java
@@ -25,14 +25,6 @@ public enum ErrorCode {
   AUCTION_CANNOT_DELETE(HttpStatus.CONFLICT, "현재 상태의 경매는 삭제할 수 없습니다."),
   AUCTION_CANNOT_CANCEL(HttpStatus.CONFLICT, "현재 상태의 경매는 취소할 수 없습니다."),
 
-
-  // 태그
-  TAG_EMPTY(HttpStatus.BAD_REQUEST, "태그에 빈 문자열을 등록할 수 없습니다."),
-  TAG_TOO_LONG(HttpStatus.BAD_REQUEST, "태그는 최대 글자 수를 초과했습니다."),
-  TAG_INVALID_CHAR(HttpStatus.BAD_REQUEST, "태그는 한글, 영어, 숫자만 사용할 수 있습니다."),
-  TAG_CONTAINS_SPACE(HttpStatus.BAD_REQUEST, "태그에 공백을 포함할 수 없습니다."),
-
-
   // 그 외
   UNKNOWN_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "알 수 없는 에러")
   ;
